app/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.32
python-dotenv==1.0.1
pytz==2025.2
pandas==2.2.3
openpyxl==3.1.2
XlsxWriter==3.2.3
selenium==4.23.1
